+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2024-06-11
+### Added
+- A markdown Preview accessible with `Ctrl + P` or the new `Preview` button on the bottom left
+- Theme switching between light and dark, will be automatically saved in the settings
+- Minimal width and height set
+- Upon creating a new note, that one will be selected
+
+### Changed
+- Upon creating a new note, that one will be selected
+- Default settings now includes a `theme` key
+- Colors updated using Tailwind to be switched to the selected theme
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- N/A
+
 ## [1.1.0] - 2024-06-09
 ### Added
 - Descriptions in the command palette can be toggled (on/off)